# ergosense-app/server/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ErgoSense API",
    "version": "1.0.0",
    "description": "API REST multi-tenant para SST, ergonomia, GRO, PGR, AET, psicossocial, eSocial e compliance."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Gateway relativo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "POST /api/auth/login",
        "security": [],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "POST /api/auth/refresh",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "POST /api/auth/forgot-password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password/preview": {
      "get": {
        "tags": [
          "Autenticação"
        ],
        "summary": "GET /api/auth/reset-password/preview",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "POST /api/auth/reset-password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/access-requests": {
      "post": {
        "tags": [
          "access-requests"
        ],
        "summary": "POST /api/access-requests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/public/support-contact": {
      "get": {
        "tags": [
          "Público"
        ],
        "summary": "GET /api/public/support-contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "post": {
        "tags": [
          "Público"
        ],
        "summary": "POST /api/public/support-contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/denuncias/public": {
      "post": {
        "tags": [
          "denuncias"
        ],
        "summary": "POST /api/denuncias/public",
        "security": [],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/auth/mfa/verify": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "POST /api/auth/mfa/verify",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/openapi.json": {
      "get": {
        "tags": [
          "openapi.json"
        ],
        "summary": "GET /api/openapi.json",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/docs": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "GET /api/docs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
...
    "/api/public/tenant-request": {
      "post": {
        "tags": [
          "Público"
        ],
        "summary": "POST /api/public/tenant-request",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/public/plans": {
      "get": {
        "tags": [
          "Público"
        ],
        "summary": "GET /api/public/plans",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/auth/activate-account/preview": {
      "get": {
        "tags": [
          "Autenticação"
        ],
        "summary": "GET /api/auth/activate-account/preview",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/auth/activate-account": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "POST /api/auth/activate-account",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}